+++ idp_engine/Interpret.py
@@ -47,7 +47,7 @@ from .Expression import (SymbolExpr, Expression, Constructor, AQuantification,
                     AImplication, AConjunction,  AEquivalence, AAggregate,
                     AComparison, AUnary, AppliedSymbol, UnappliedSymbol, Number,
                     Variable, TRUE)
-from .utils import BOOL, RESERVED_SYMBOLS, SYMBOL, OrderedSet
+from .utils import BOOL, RESERVED_SYMBOLS, SYMBOL, OrderedSet, DEFAULT
 
 
 # class Extern  ###########################################################
@@ -131,7 +131,7 @@ Rule.interpret = interpret
 # class SymbolInterpretation  ###########################################################
 
 def interpret(self, problem):
-    status = (Status.STRUCTURE if self.block.name != 'default' else
+    status = (Status.STRUCTURE if self.block.name != DEFAULT else
                 Status.GIVEN)
     if self.is_type_enumeration:
         self.enumeration.interpret(problem)
@@ -147,9 +147,8 @@ def interpret(self, problem):
                 or problem.assignments[expr.code].status == Status.UNKNOWN,
                 f"Duplicate entry in structure for '{self.name}': {str(expr)}")
             e = problem.assignments.assert_(expr, value, status, False)
-            if (self.block.name == 'default'
-                and type(self.enumeration) == FunctionEnum
-                and type(self.symbol.decl.out.decl) == TypeDeclaration):
+            if (status == Status.GIVEN  # for proper display in IC
+                and type(self.enumeration) == FunctionEnum):
                 problem.assignments.assert_(e.formula(), TRUE, status, False)
         if self.default is not None:
             for code, expr in self.symbol.decl.instances.items():
@@ -157,6 +156,10 @@ def interpret(self, problem):
                     or problem.assignments[code].status != status):
                     e = problem.assignments.assert_(expr, self.default, status,
                                                 False)
+                    if (status == Status.GIVEN  # for proper display in IC
+                        and self.default.type != BOOL):
+                        problem.assignments.assert_(e.formula(), TRUE, status,
+                                                    False)
 SymbolInterpretation.interpret = interpret