+++ idp_engine/Problem.py
@@ -216,7 +216,8 @@ class Problem(object):
             self._formula = AConjunction.make(
                 '∧',
                 [a.formula() for a in self.assignments.values()
-                 if a.value is not None]
+                 if a.value is not None
+                 and a.status not in [Status.CONSEQUENCE, Status.ENV_CONSQ]]
                 + [s for s in self.constraints]
                 + [c for c in self.co_constraints]
                 + [s for s in self.def_constraints.values()]
@@ -227,7 +228,8 @@ class Problem(object):
     def _todo(self):
         return OrderedSet(
             a.sentence for a in self.assignments.values()
-            if a.value is None
+            if a.status not in [Status.GIVEN, Status.STRUCTURE,
+                                 Status.UNIVERSAL, Status.ENV_UNIV]
             and a.symbol_decl is not None
             and (not a.sentence.is_reified() or self.extended))
 
@@ -235,20 +237,15 @@ class Problem(object):
         """ returns Assignments from model in solver """
         ass = self.assignments.copy()
         for q in todo:
-            if not q.is_reified():
-                val1 = solver.model().eval(
-                    q.translate(),
-                    model_completion=complete)
-            elif self.extended:
+            val1 = None
+            if not q.is_reified() or self.extended:
+                # evaluating q.translate() directly fails the pipeline on arithmetic/forall.idp
                 solver.push()  # in case todo contains complex formula
                 solver.add(q.reified() == q.translate())
                 res1 = solver.check()
                 if res1 == sat:
-                    val1 = solver.model().eval(
-                        q.reified(),
-                        model_completion=complete)
-                else:
-                    val1 = None  # dead code
+                    val1 = solver.model().eval(q.reified(),
+                                               model_completion=complete)
                 solver.pop()
             if val1 is not None and str(val1) != str(q.translate()):  # otherwise, unknown
                 val = str_to_IDP(q, str(val1))
@@ -337,8 +334,8 @@ class Problem(object):
         result = solver.check()
         if result == sat:
             for q in todo:
-                solver.push()  # in case todo contains complex formula
-                solver.add(q.reified() == q.translate())
+                solver.push()  #  faster (~3%) with push than without
+                solver.add(q.reified() == q.translate())  # in case todo contains complex formula
                 res1 = solver.check()
                 if res1 == sat:
                     val1 = solver.model().eval(q.reified())
@@ -353,6 +350,8 @@ class Problem(object):
                             yield self.assignments.assert_(q, val, tag, True)
                         elif res2 == unknown:
                             res1 = unknown
+                        else:  # reset the value
+                            self.assignments.assert_(q, None, Status.UNKNOWN, False)
                 solver.pop()
                 if res1 == unknown:
                     # yield(f"Unknown: {str(q)}")
@@ -440,7 +439,12 @@ class Problem(object):
         return (facts, laws)
 
     def simplify(self):
-        """ simplify constraints using known assignments """
+        """ returns a simpler copy of the Problem, using known assignments
+
+        Assignments obtained by propagation become fixed constraints.
+        """
+
+        self = self.copy()
 
         # annotate self.constraints with questions
         for e in self.constraints:
@@ -451,6 +455,10 @@ class Problem(object):
         for ass in self.assignments.values():
             old, new = ass.sentence, ass.value
             if new is not None:
+                # convert consequences to Universal
+                ass.status = (Status.UNIVERSAL if ass.status == Status.CONSEQUENCE else
+                              Status.ENV_UNIV if ass.status == Status.ENV_CONSQ else
+                              ass.status)
                 # simplify constraints
                 new_constraints: List[Expression] = []
                 for constraint in self.constraints: