+++ idp_engine/utils.py
@@ -23,7 +23,17 @@ from collections import ChainMap, Iterable
 from json import JSONEncoder
 import time
 
+
+"""
+    Global Parameters:
+"""
+
+CO_CONSTR_RECURSION_DEPTH = 3
 MAX_QUANTIFIER_EXPANSION = 20
+
+"""
+    String constants
+"""
 
 NEWL = "\n"
 indented = "\n  "
@@ -40,11 +50,7 @@ OUTPUT_DOMAIN = "output_domain"
 RESERVED_SYMBOLS = [BOOL, INT, REAL, DATE, SYMBOL,
                     RELEVANT, ARITY, INPUT_DOMAIN, OUTPUT_DOMAIN]
 
-"""
-    Parameters:
-"""
-
-CO_CONSTR_RECURSION_DEPTH = 3
+DEFAULT = "default"
 
 """ Module that monkey-patches json module when it's imported so
 JSONEncoder.default() automatically checks for a special "to_json()"