+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Contributors (alphabetical): Matthias van der hallen, Pierre Carbonnelle.
 * fix #90: error in `sum` aggregate
 
 ### Interactive Consultant
+* Merge #85: improve performance of IC (better use of cache)
+* fix #75: default structure not working in Registration application
 * fix #89: `constructed from` types create error in IC
 * fix #93: incorrect parsing of `main()` block in IDE